requirements.txt
tensorflow
opencv-python
pillow
streamlit
matplotlib
textblob
scikit-learn
numpy
pandas